# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rag-jev",
    "description": "Select useful context from existing retrieval results. No ingestion required.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/config": {
      "get": {
        "summary": "Config",
        "operationId": "config_v1_config_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Config V1 Config Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recordings": {
      "get": {
        "summary": "Recordings",
        "operationId": "recordings_v1_recordings_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Recordings V1 Recordings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/examples": {
      "get": {
        "summary": "Examples",
        "operationId": "examples_v1_examples_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Examples V1 Examples Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fixtures/{case_id}": {
      "get": {
        "summary": "Fixture",
        "operationId": "fixture_v1_fixtures__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs": {
      "post": {
        "summary": "Runs",
        "operationId": "runs_v1_runs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/replay": {
      "post": {
        "summary": "Replay",
        "operationId": "replay_v1_replay_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/import": {
      "post": {
        "summary": "Import Run",
        "operationId": "import_run_v1_import_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayBundle"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportedRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/compare": {
      "post": {
        "summary": "Compare",
        "operationId": "compare_v1_compare_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/select": {
      "post": {
        "summary": "Select",
        "operationId": "select_v1_select_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rankings": {
      "post": {
        "summary": "Rankings",
        "operationId": "rankings_v1_rankings_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/compare-rankings": {
      "post": {
        "summary": "Ranking Answers",
        "operationId": "ranking_answers_v1_compare_rankings_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Comparison": {
        "properties": {
          "input_hash": {
            "type": "string",
            "title": "Input Hash"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "baseline": {
            "$ref": "#/components/schemas/Answer"
          },
          "selected": {
            "$ref": "#/components/schemas/Answer"
          },
          "baseline_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Baseline Ids"
          },
          "selected_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Ids"
          },
          "selection_source": {
            "type": "string",
            "enum": [
              "live",
              "fixture"
            ],
            "title": "Selection Source"
          },
          "baseline_stage_ms": {
            "type": "number",
            "title": "Baseline Stage Ms"
          },
          "selected_stage_ms": {
            "type": "number",
            "title": "Selected Stage Ms"
          },
          "selection_original_ms": {
            "type": "number",
            "title": "Selection Original Ms"
          },
          "baseline_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Cost Usd"
          },
          "selected_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Cost Usd"
          },
          "selection_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selection Cost Usd"
          },
          "cost_basis": {
            "type": "string",
            "title": "Cost Basis",
            "default": "Estimate from configured per-million rates; excludes retrieval and caching discounts"
          },
          "timing_basis": {
            "type": "string",
            "title": "Timing Basis",
            "default": "Generation measured now; selected stage adds original scoring time. Retrieval excluded."
          },
          "citation_check": {
            "type": "string",
            "title": "Citation Check",
            "default": "Citation IDs checked; factual support and answer correctness are not graded"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "input_hash",
          "policy_version",
          "baseline",
          "selected",
          "baseline_ids",
          "selected_ids",
          "selection_source",
          "baseline_stage_ms",
          "selected_stage_ms",
          "selection_original_ms",
          "baseline_cost_usd",
          "selected_cost_usd",
          "selection_cost_usd"
        ],
        "title": "Comparison"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportedRun": {
        "properties": {
          "view": {
            "$ref": "#/components/schemas/RunView"
          },
          "comparison": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Comparison"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "view"
        ],
        "title": "ImportedRun"
      },
...
      "RankingRequest": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/ScoredRun"
          },
          "top_n": {
            "type": "integer",
            "maximum": 256.0,
            "minimum": 1.0,
            "title": "Top N",
            "default": 10
          },
          "max_context_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Context Tokens"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "record"
        ],
        "title": "RankingRequest"
      },
      "ReplayBundle": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/ScoredRun"
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          },
          "comparison": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Comparison"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "record",
          "policy"
        ],
        "title": "ReplayBundle"
      },
      "ReplayRequest": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/ScoredRun"
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "record",
          "policy"
        ],
        "title": "ReplayRequest"
      },
      "RunView": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/ScoredRun"
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          },
          "selection": {
            "$ref": "#/components/schemas/SelectionResult"
          },
          "baseline_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Baseline Ids"
          },
          "baseline_context_tokens": {
            "type": "integer",
            "title": "Baseline Context Tokens"
          },
          "selected_context_tokens": {
            "type": "integer",
            "title": "Selected Context Tokens"
          },
          "token_estimator": {
            "type": "string",
            "title": "Token Estimator",
            "default": "cl100k_base estimate; actual prompt usage appears after generation"
          },
          "evidence_recall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Recall"
          },
          "relevant_dropped": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevant Dropped"
          },
          "baseline_evidence_recall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Evidence Recall"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "record",
          "policy",
          "selection",
          "baseline_ids",
          "baseline_context_tokens",
          "selected_context_tokens",
          "evidence_recall",
          "relevant_dropped",
          "baseline_evidence_recall"
        ],
        "title": "RunView"
      },
...
      "StudyRequest": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/SelectRequest"
          },
          "relevant_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevant Ids"
          },
          "baseline_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request"
        ],
        "title": "StudyRequest"
      },
...
    }
  }
}